requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
crispy-bootstrap5==0.7
Django==4.1.7
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
